rewrite and add a bunch of features
this is a bit bigger commit than I wanted but, one change lead to
another and now then everything got tangled up.

the changes are

1. instead of using a list of just bad-extension id's, we use the
current-list-meta.csv which has some info about the extensions it finds,
i.e. name and link to more info, we now display that info if we find
a bad extension

2. instead of using a separate checksum file, we just check that the
header matches the meta file, that removes the `shasum` dependency which
didn't work on OS X Mojave as that shasum build lacks the -p flag

3. replaced the OS checking to decide which paths to check, with some
logic that just scans all known paths that exists. This should make
it easier to add other OS's and Chromium browsers which might use
the Chrome web store

4. instead of downloading the list of bad extensions every time, it now
downloads current-list-meta.csv into /tmp/bad-chrome-extensions.csv and
use `curl -z` to avoid downloading it again if it hasn't changed.